Refactoring code
changed the README to use keystone
moved the unicode helpers to the tools package
