+++ snappy/generators.py
@@ -3,7 +3,7 @@ import math
 from random import choice
 
 from hashlib import md5
-from unicode_helpers import UNICODE_BLOCKS, BLOCK_NAMES
+from snappy.tools.unicode_helpers import (UNICODE_BLOCKS, BLOCK_NAMES)
 from snappy.common.datasets import DatasetList
 from swift_composite import ObjectStorageComposite
 from configs.swift_config import ObjectStorageAPIConfig
