+++ snappy/swift_behaviors.py
@@ -10,7 +10,7 @@ from random import choice
 from StringIO import StringIO
 from time import sleep, time
 
-from unicode_helpers import UNICODE_BLOCKS, BLOCK_NAMES
+from snappy.tools.unicode_helpers import (UNICODE_BLOCKS, BLOCK_NAMES)
 from snappy.tools.md5hash import get_md5_hash
 from os_client import ObjectStorageAPIClient
 from constants import Constants
