+++ snappy/tools/unicode_helpers.py
